Use different events APIs based on API version
Rather than always trying the new endpoint and then using the
old endpoint if we get a 404.

[#64525814]

+++ src/cf/commands/application/events_test.go
@@ -69,7 +69,7 @@ var _ = Describe("events command", func() {
 
 		runCommand("my-app")
 
-		Expect(eventsRepo.RecentEventsArgs.Limit).To(Equal(uint(50)))
+		Expect(eventsRepo.RecentEventsArgs.Limit).To(Equal(uint64(50)))
 		Expect(eventsRepo.RecentEventsArgs.AppGuid).To(Equal("my-app-guid"))
 		testassert.SliceContains(ui.Outputs, testassert.Lines{
 			{"Getting events for app", "my-app", "my-org", "my-space", "my-user"},
